Move microscope.testsuite.devel into microscope.testsuite.devices
* microscope/testsuite/devel.py, microscope/testsuite/devices.py: all
  test devices are in microscope.testsuite.devices so move everything
  from microscope.testsuite.devel, all test devices, in there.
